updated track reconstruction, added sg vs bg classifier
